[runtimebroker.ps1] Script to launch BugId on Windows 10 (10.0.17134.81)
As defined in this thread https://github.com/SkyLined/BugId/issues/80 there is a bug in Windows or windbg/cdb that makes impossible to debug edge. The issue is around the process RuntimeBroker.exe, sometimes when BugId starts, this process is suspended. This script will start a monitoring thread that will track the process status. When it's suspended BugId is stopped and the affected processes are killed. Then BugId is restarted and RuntimeBroker.exe is checked again. Once everything works fine, control thread exits to avoid unnecessary BugId restarting.